Move the LoggableError type to the error module
It doesn't make so much sense to have it in the logger module since it
doesn't depend on the specific logger implementation.
Additionally, it enables a more strict dependency policy.
