+++ src/util.rs
@@ -22,7 +22,7 @@ use libgitdit::{Issue, RepositoryExt};
 use error::*;
 use error::ErrorKind as EK;
 use gitext::RemotePriorization;
-use system::{Abortable, IteratorExt, LoggableError};
+use system::{Abortable, IteratorExt};
 
 /// Open the DIT repo
 ///
